refactor: ChatServer 빌드 경고 캐스트로 정리
- GetCharacterSize 반환 시 size_t → int 명시 캐스트 (C4267)
- _MaxUser 비교 시 int 캐스트 (C4018)
- main.cpp CPU·Network 통계 표시용 (int) 캐스트 4건 (C4244)

+++ Examples/ChatServer/Content.cpp
@@ -30,7 +30,7 @@ Chat_Server::Chat_Server(int MaxUser, bool HeartBeatFlag)
 
 bool Chat_Server::OnConnectionRequest()
 {
-	if (GetCharacterSize() < _MaxUser) return true;
+	if (GetCharacterSize() < (int)_MaxUser) return true;
 	else return false;
 }
